Copying of all binary builds to output directory Signed-off-by: Sergey Matveev <stargrave@stargrave.org>

+++ build/build.sh
@@ -5,6 +5,7 @@ pushd $WORK_DIR; WORK_DIR=`pwd`; popd
 TARGET_DIR="$WORK_DIR"/openwrt/trunk
 OUTPUT_DIR="$WORK_DIR"/output/`date "+%Y-%m-%dT%H:%M"`
 DL_DIR="$TARGET_DIR"/dl
+BIN_DIR="$TARGET_DIR"/bin/x86
 
 mmake()
 {
@@ -70,7 +71,7 @@ update_openwrt_config()
 
 create_dl_directory()
 {
-	[ -L "$DL_DIR" ] && [ `readlink "$DL_DIR"` = "$DL_PATH" ] && return 0
+	[ -L "$DL_DIR" ] && [ `readlink "$DL_DIR"` = "$DL_PATH" ] && return 0 || true
 	if [ -d "$DL_DIR" ]; then
 		rm -fr "$DL_DIR"
 	else
@@ -94,6 +95,12 @@ perform_building()
 	mmake -j V=99 >"$OUTPUT_DIR"/output.log 2>&1
 }
 
+copy_bins()
+{
+	[ -d "$BIN_DIR" ] || return 0
+	cp -a "$BIN_DIR"/* "$OUTPUT_DIR"/
+}
+
 check_openwrt_existence
 update_feeds_configuration
 update_openwrt_config
@@ -101,3 +108,4 @@ create_dl_directory
 create_output_directory
 perform_cleaning
 perform_building
+copy_bins
